[#53]Feat: Long rate -> Float rate 타입 변환

+++ src/main/java/org/nbc/account/trollo/domain/card/dto/response/CardReadResponseDto.java
@@ -13,7 +13,7 @@ public record CardReadResponseDto(
     String color,
     LocalDateTime deadline,
     List<CheckListRequestDto> checkList,
-    Long rate
+    Float rate
 ) {
 
 }
